refactor: auth context provider and config
