+++ src/config/index.ts
@@ -2,6 +2,7 @@ type ApiConfigType = {
   apiBasePath: string
   oidcDiscoveryUrl: string
   oidcClientId: string
+  redirectUri: string
 }
 
 const apiConfig: ApiConfigType = {
@@ -9,7 +10,8 @@ const apiConfig: ApiConfigType = {
   oidcDiscoveryUrl:
     process.env.OPENID_CONNECT_DISCOVERY_URL ||
     'http://localhost:8080/auth/realms/tdp_server_dev/.well-known/openid-configuration',
-  oidcClientId: process.env.OPENID_CLIENT_ID || 'tdp_server_id',
+  oidcClientId: process.env.OPENID_CLIENT_ID || 'tdp_auth',
+  redirectUri: 'http://localhost:3000/login',
 }
 
 export default apiConfig
