+++ src/contexts/index.ts
@@ -1,1 +1,2 @@
+export * from './AuthContextProvider'
 export * from './TdpClientContext'
